refactor: extract project retrieval into client component
